First Release - Storage Depots added in Ados/Deniran with NPCs to rent
Storage Depot buildings have now been added in the city of Ados and
Deniran, where players may go inside, talk with an npc to rent one of
the 6 rooms full of chests. I've added similar classes to the BuyHouse
classes but for StorageRoom and added "storage room key" to be able to
access the rooms. StorageRoomPortals have been used to go in and out
ofthe storage rooms.