Ados_Ship Zone Created in Ados as 0_ados_ship_e2
Following the example of 0_athor_ship_w2, we created the new zone in the
coordinates of a non existing zone just right of 0_ados_ocean_e and just
below 0_amazon_island_se(use the atlas to figure out if needed).
The ados ship zone is essentially the same as the athor ship, but has
different NPCs (so we recreate the same ferry behaviour but going to
different places).
Conceptually the coordinates of the new ship make no sense because its
far away from the ConveyerNPC, but this realistically doesn't matter.